clean up + reorganize skill components
