--- frontend/common/src/components/skills/SearchBar/index.ts
@@ -1,5 +1,0 @@
-import SearchBar from "./SearchBar";
-import type { SearchBarProps } from "./SearchBar";
-
-export default SearchBar;
-export type { SearchBarProps };
