--- frontend/common/src/components/skills/SkillFamilyPicker/index.ts
@@ -1,5 +1,0 @@
-import SkillFamilyPicker from "./SkillFamilyPicker";
-import type { SkillFamilyPickerProps } from "./SkillFamilyPicker";
-
-export default SkillFamilyPicker;
-export type { SkillFamilyPickerProps };
