--- frontend/common/src/components/skills/SkillPicker/deprecated/index.ts
@@ -1,3 +1,0 @@
-import SkillPicker from "./SkillPicker";
-
-export default SkillPicker;
