--- frontend/common/src/components/skills/SkillResults/index.ts
@@ -1,5 +1,0 @@
-import SkillResults from "./SkillResults";
-import type { SkillResultsProps } from "./SkillResults";
-
-export default SkillResults;
-export type { SkillResultsProps };
