+++ frontend/talentsearch/src/js/components/reviewMyApplication/ReviewMyApplicationPage.tsx
@@ -5,7 +5,7 @@ import { ArrowSmallRightIcon } from "@heroicons/react/24/solid";
 import UserProfile from "@common/components/UserProfile";
 import Pending from "@common/components/Pending";
 import NotFound from "@common/components/NotFound";
-import MissingSkills from "@common/components/skills/MissingSkills";
+import MissingSkills from "@common/components/MissingSkills";
 import ExperienceSection from "@common/components/UserProfile/ExperienceSection";
 import Well from "@common/components/Well";
 import { commonMessages, navigationMessages } from "@common/messages";
