--- frontend/talentsearch/src/js/components/skills/AddSkillsToExperience/index.ts
@@ -1,2 +1,0 @@
-// eslint-disable-next-line no-restricted-exports
-export { default, AddSkillsToExperienceProps } from "./AddSkillsToExperience";
